django_projects/blogsite2.0/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.18.64
botocore==1.21.64
Django==3.2.8
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-js-asset==1.2.2
django-storages==1.12.2
django-taggit==3.0.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
